challenge1 / clean arch
Handled all error properly and reorganized handlers.
