+++ cmd/api/main.go
@@ -10,7 +10,6 @@ import (
 const port = ":8080"
 
 func main() {
-
 	hh := handlers.NewHomeHandler()
 	rh := handlers.NewReadHandler()
 
@@ -18,8 +17,7 @@ func main() {
 	http.HandleFunc("/read", rh.ReadLocalCsv)
 
 	log.Println("Listening on port", port[1:])
-	err := http.ListenAndServe(port, nil)
-	if err == nil {
+	if err := http.ListenAndServe(port, nil); err != nil {
 		log.Fatal(err)
 	}
 }
