+++ pkg/handlers/home.endpoint.go
@@ -7,17 +7,27 @@ import (
 	"github.com/valp0/academy-go-q32021/pkg/utils"
 )
 
+type homeHandler struct{}
+
+type IHomeHandler interface {
+	Home(w http.ResponseWriter, r *http.Request)
+}
+
+// Will return homeHandler struct.
+func NewHomeHandler() IHomeHandler {
+	return homeHandler{}
+}
+
 // The entry point handler.
 func (hh homeHandler) Home(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 
-	jRes, _ := utils.JsonResponse("At this stage, the only available endpoint is /read.")
+	jRes := utils.JsonResponse("At this stage, the only available endpoint is /read.")
 	prettified := utils.Prettify(jRes)
 
 	w.WriteHeader(http.StatusOK)
-	_, err := fmt.Fprintf(w, prettified)
-	if err != nil {
-		fmt.Println(err)
-		return
+	b, err := fmt.Fprintf(w, prettified)
+	if err != nil || b < 1 {
+		fmt.Println("Couldn't write response bytes to http.ResponseWriter,", err)
 	}
 }
